+++ client/app/types/props/organisms/GranLinkList.ts
